pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vouchersnap"
version = "0.1.0"
description = "Herbarium specimen photo manager - scan QR codes and upload to iNaturalist"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "VoucherSnap Contributors"}
]
keywords = ["herbarium", "iNaturalist", "mycology", "specimen", "QR code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "pyinaturalist>=0.19",
    "Pillow>=10.0",
    "pillow-heif>=0.10",
    "pyzbar>=0.1.9",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
vouchersnap = "vouchersnap.cli:cli"

[project.urls]
Homepage = "https://github.com/vouchersnap/vouchersnap"
Repository = "https://github.com/vouchersnap/vouchersnap"

[tool.setuptools.packages.find]
where = ["src"]
